fix: getMask should return filled ROI
