+++ src/roi/Roi.ts
@@ -77,7 +77,7 @@ export class Roi {
    * @param options - Get Mask options
    * @returns The ROI mask.
    */
-  public getMask(options: GetMaskOptions = { kind: 'contour' }): Mask {
+  public getMask(options?: GetMaskOptions): Mask {
     return getMask(this, options);
   }
